+++ proofs/src/plonk/prover.rs
@@ -27,8 +27,8 @@ use crate::{
         traces::ProverTrace, trash,
     },
     poly::{
-        batch_invert_rational, commitment::PolynomialCommitmentScheme, Coeff, CommitmentLabel,
-        EvaluationDomain, ExtendedLagrangeCoeff, LagrangeCoeff, Polynomial,
+        batch_invert_rational, commitment::PolynomialCommitmentScheme, Coeff, EvaluationDomain,
+        ExtendedLagrangeCoeff, LagrangeCoeff, Polynomial, PolynomialLabel,
         PolynomialRepresentation, ProverQuery, Rotation,
     },
     transcript::{Hashable, Sampleable, Transcript},
@@ -54,7 +54,7 @@ where
     for (poly_eval, value) in poly.iter_mut().zip(instances.iter()) {
         *poly_eval = *value;
     }
-    CS::commit(params, &poly, CommitmentLabel::Instance(0))
+    CS::commit(params, &poly, PolynomialLabel::Instance(0))
 }
 
 /// This computes a proof trace for the provided `circuit` when given the
@@ -198,7 +198,11 @@ where
                         .par_iter()
                         .map(|h| {
                             let h_poly = domain.lagrange_from_vec(h.clone());
-                            CS::commit(params, &h_poly, CommitmentLabel::NoLabel)
+                            CS::commit(
+                                params,
+                                &h_poly,
+                                PolynomialLabel::LogupHelper(c.name.clone()),
+                            )
                         })
                         .collect()
                 })
@@ -503,7 +507,11 @@ where
             }
 
             if is_committed_instance {
-                transcript.common(&CS::commit(params, &poly, CommitmentLabel::NoLabel))?;
+                transcript.common(&CS::commit(
+                    params,
+                    &poly,
+                    PolynomialLabel::CommittedInstance(i),
+                ))?;
             }
 
             Ok(poly)
@@ -605,7 +613,7 @@ where
     let advice_commitments: Vec<_> = advice_values
         .par_iter()
         .enumerate()
-        .map(|(i, poly)| CS::commit(params, poly, CommitmentLabel::Advice(i)))
+        .map(|(i, poly)| CS::commit(params, poly, PolynomialLabel::Advice(i)))
         .collect();
 
     for commitment in &advice_commitments {
@@ -721,7 +729,7 @@ where
             values: h_poly,
             _marker: std::marker::PhantomData,
         };
-        let h_com = CS::commit(params, &h_poly, CommitmentLabel::NoLabel);
+        let h_com = CS::commit(params, &h_poly, PolynomialLabel::Quotient);
         transcript.write(&h_com)?;
         Ok(vec![h_poly])
     }
@@ -743,7 +751,8 @@ where
         // Compute commitment to each limb (parallel MSMs).
         let h_commitments: Vec<_> = h_limbs
             .par_iter()
-            .map(|h_piece| CS::commit(params, h_piece, CommitmentLabel::NoLabel))
+            .enumerate()
+            .map(|(i, h_piece)| CS::commit(params, h_piece, PolynomialLabel::QuotientPiece(i)))
             .collect();
 
         // Write each limb commitment to the transcript in order.